requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
google-generativeai
python-dotenv
pillow